Rename downsyncinventory class to takeInventorySnapshots
